requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
# sqlite3-utils>=3.35.0  # Built into Python as sqlite3
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.31.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Telegram Bot (for bot approach)
python-telegram-bot>=20.7

# Telegram User Client (for personal account approach)
telethon>=1.34.0
cryptg>=0.4.0

# Exchange APIs (optional - install when ready for live trading)
# python-binance>=1.0.0
# cbpro>=1.1.4
# krakenex>=2.1.0